+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 # Enhancements
 * Added textures for 4 rendertargets (camera, water reflection/refraction, full framebuffer). These give a nice preview in Hammer.
+* Add ability to randomise output delay to `comp_relay` and `comp_adv_output`.
 
 
 --------------------